chore(lint): adopt golangci-lint, with the tree brought to what it checks
CI's reusable lint workflow runs golangci-lint whenever a configuration
file exists, and none did, so it ran gofmt and go vet and nothing else.
.golangci.yml now enables the linters the tree meets plus the ones cheap to
keep meeting — errcheck, staticcheck, unused, ineffassign, unconvert,
misspell, errorlint, nilerr, noctx, prealloc, bodyclose, gosec — with
every exclusion stated beside its reason. `make lint` runs the same pinned
version, and `make ci` includes it.

What bringing the tree to it found, beyond style: two tests that compared
an expression with itself and so could never fail (a spec-hash and a
connection-stamp determinism check); two written files whose close was
never checked, where a full disk reports itself (a fetched tool, an
extracted one); a kubectl call composing a failure hint that could hang on
a broken kubeconfig, now bounded; a dozen deliberate nil-error returns and
weak-hash imports annotated with the reason a reader would otherwise
reconstruct; read-side closes made explicit about ignoring their error;
and invisible characters in test strings written as the escapes they are.

+++ builtin/internal/gitclone/gitclone.go
@@ -71,7 +71,7 @@ type cappedStorage struct {
 }
 
 func newCappedStorage() *cappedStorage {
-	remaining := int64(maxObjectBytes)
+	remaining := maxObjectBytes
 	return &cappedStorage{Storer: memory.NewStorage(), remaining: &remaining}
 }
